Water-Management-System
Water Quality Management System for the Smart City is the system based
on the internet of things concept where the system is composed of the
sensors interconnected to each other.
The System is taking readings and inputs from the sensors measuring the
water quality of each household community or home situated in the smart
city.
When a reading is found abnormal in the system an alert or work request
is generated to the surveyor so that he can find the exact issue by
visiting the location.
After finding the type of issue the request can be forwarded to the
maintenance department to fix the issue.
Also the residents can also login into the application and enter the
vital signs details so that it can be analyzed for the health issues
caused by the bad quality of water or water supply defect.
The sensor data can be used by many Enterprises or Departments for
research and analysis. The Employees can communicate to each other
through work requests.